:alembic: Experiment with lots of concurrent connections to get things
Given how terrible the HackerNews API is, and given that it seems to need a
metric fuckton of API calls to get any useful data, and given that being
kind and doing them in serial is slow as; let's try and do as many
concurrent connections as possible.

So far this seems way faster (no surprise) while also seeming like total
overkill.

But I think I'll carry on with this approach for the moment.