:construction: WiP comment loading
The core code for loading the comments is in here, and I'm still riffing on
how I'm going to show the comments and let the user navigate them. Still to
be fully decided.

Note that right now all it does is load the top level of comments, part of
what I want to decide is how I do the different levels.