:zap: Tidy up how the first items gets highlighted
Some change in the way that selected options gets handled, that came in with
0.48.x (https://github.com/Textualize/textual/pull/3912 I believe), has
broken the initial highlight in this application. I've noticed something odd
like this in tinboard too but can't pin it down. There seems to be some odd
"hidden" state for the highlight in some circumstance I can't quite track
down yet.

This change however makes clear the intent I have for this application, and
also solves the problem.

I'll see if I can isolate the issue soon.