:sparkles: By default, load as-yet-unread tabs in the background
The idea here is that, once the first tab is loaded and the user is reading
the headlines, etc, the next as-yet-unloaded tab starts to load its stories
in the background; when it has finished the next is asked to load, and so
on.

In other words, by the time you've finished looking over "Top", all the
other tabs should already be loaded and ready, making the whole thing feel
faster.

Set this as an option in the application configuration so it can be turned
off.